Rename file secret_service_store.rs
