+++ linux/user-daemon/src/stores/mod.rs
@@ -2,7 +2,7 @@ use u2f_core::{ApplicationKey, Counter};
 
 pub(crate) mod file_store;
 pub(crate) mod file_store_v2;
-pub(crate) mod secret_service;
+pub(crate) mod secret_service_store;
 
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct Secret {
