+++ linux/user-daemon/src/stores/secret_service_store.rs
